backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
databases==0.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minio==7.2.15
multidict==6.4.3
ormar==0.20.2
packaging==24.2
propcache==0.3.1
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.1
python-dotenv==1.1.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
yarl==1.20.0
